Added new `Export as HTML` option
